\name{iGasso-package}
\alias{iGasso-package}
\alias{iGasso}
\docType{package}
\title{
Statistical Tests for Genetic Association
}
\description{
iGasso is a collection of statistical tests developed for genetic association studies. So far it contains functions for analysis of multiple phenotypes and analysis of rare variants. It will keep growing as more tests are developed. Use \code{?iGasso} to see an introduction.
}
\details{
\tabular{ll}{
Package: \tab iGasso\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-01-07\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
Functions for various tests have .test as their extension name. 
}
\author{
Kai Wang \code{<kai-wang@uiowa.edu>}
}
\references{
Anscombe F.J. (1948) The transformation of Poisson, binomial and negative-binomial data. \emph{Biometrika} \bold{35(3/4)}, 246--254.

Freeman, M. F., Tukey, J. W. (1950) Transformations related to the angular and the square root. \emph{The Annals of Mathematical Statistics} \bold{21(4)}, 607--611.

Wang, K. (2012) An application of the proportional odds model to genetic association studies. Submitted.

Wang, K., Fingert, J. (2012) Statistical tests for detecting rare variants using variance-stabilizing transformations. Submitted. 

Zar, J.H. (1999) \emph{Biostatistical Analysis, 4th ed.}, New Jersey:Prentice-Hall, Inc.
}

%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
%%\keyword{ package }
\seealso{
%%~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~simple examples of the most important functions ~~
}
